Move most protein-related registrations to ProteinModule. CR feedback.
